refactor: move Tile into mod tiles
